User commit

+++ src/main/java/com/example/demo/repository/FertilizerRepository.java
@@ -1,11 +1,12 @@
 package com.example.demo.repository;
 
-import com.example.demo.entity.Fertilizer;
-import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+import com.example.demo.entity.Fertilizer;
 
-public interface FertilizerRepository extends JpaRepository<Fertilizer, Long> {
+public interface FertilizerRepository extends JpaRepository<Fertilizer, Integer> {
 
     List<Fertilizer> findByRecommendedForCropsIn(List<String> cropNames);
 
 }
+
